First commit in a long time.
Although I put this in git last July, I forgot about it.  I've been updating
the log pretty much several times a day and have even been sharing it with
people who have asked, but I haven't yet cleaned this up enough to feel
comfortable making it a public git repository.  On the other hand, sometimes
we benefit from doing things we're not comfortable with, so ...