Includes calories up through today's lunch.
I got curious about what I weighed a year ago and looked up the first Sunday
of Balloon fiesta and found comments about a SAR mission.  It is incredible
how time flies, because it feels like that mission was years ago because so
much has happened since.

The biggest write-up of that mission is contained in:
http://mra.org/wp-content/uploads/2019/09/Meridian_Summer_2019.pdf

but others appear in:
https://sar-taos.org/injured-campers-rescued-near-trampas/
https://www.statesmanjournal.com/story/news/local/2019/04/05/goodfellow-student-awarded-saving-lives-during-camping-mishap/3375162002/
https://www.goodfellow.af.mil/Newsroom/Article-Display/Article/1806235/student-saves-2-lives-receives-medal/
https://www.dvidshub.net/news/316998/student-saves-2-lives-receives-medal
https://www.amarillo.com/news/20181216/beilue-that-day-on-mountain
and
https://www.33fw.af.mil/News/Article-Display/Article/1806523/student-saves-2-lives-receives-medal/