Includes through today's fruit snack.
Like usual, I didn't do a good job of recording what all I had on
Saturday after the Bosque Bigfoot 50k.  A contributing factor is that
Sophie and I went over to Michelle & Collin's to a party that was
celebrating their wedding anniversary (the 14th) and Collin's birthday
(today, the 15th).  On top of that, there was a storm rolling in and
sure enough, I had to drive back in a snowstorm.