Includes through today's late afternoon beans.
I have not gone back and tried to fill in a bunch of missing blanks.
I have photos of all the different beers we drank during Balloon
Fiesta and I have wrappers from the Tailwind packages I used during
AA50k, but I got further and further behind and so I don't know if
I'll backpatch or not.

I'm pushing this to GitHub now so that if my laptop dies I won't lose
a bunch of data.