Includes up through afternoon snack.
This commit includes a bunch of drinking and eating that I did during
the Balloon Fiesta and some of what I remember eating at the Feral Hog
50k.  Much of the Balloon Fiesta drinking is listed as "3 oz.?", but
those were typically 12 oz. bottles being shared with as many people
as were interested and around, which was typically more than four, so
3 oz. is probably high, but I don't want to kid myself either.  I
drank more than I had planned, in part because I wasn't driving (I ran
to the field from my house and then caught a lift back) and in part
because for Wednesday and Thursday, I thought the Feral Hog 50k was
going to be canceled.

Oh and when I drank a lot I also ate a ton of junk food.  So, I wound
up gaining about five pounds between Friday, September 30th and
Monday, October 10th, but that also included the Mt. Taylor 50k and
the Feral Hog 50k and I ate poorly at the aid stations and post-race
in part because I'm broke and it was free.  Oh and the Balloon Fiesta
beer was free too, due in part to Tom but especially Bill.  I owe them
next year.