includes through today's lunch, but doesn't specify which beer I had during AIBF
As is tradition, I took a picture of the bottles, so it's just a
matter of going through them and entering the data, but I'm behind on
so many things that it may be a while before I get around to it and I
might even forget to do so.

FWIW, I actually lost weight during Balloon Fiesta, but probably not
as much as I'd have lost had it not been AIBF.  I didn't get much
programming done, but I did manage to prepare and eat my food and get
Craft Poker Co.'s first tax filing done (via Clayton Financial;
thanks, Russ!).